refactor(router): share first-party picker core
